git-svn-id: svn://192.168.1.10/public/CoffeeMud@8122 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Abilities/Druid/Chant_GrowFood.java
@@ -68,12 +68,12 @@ public class Chant_GrowFood extends Chant
 		StringBuffer buf=new StringBuffer("Food types known:\n\r");
 		List<Integer> codes = RawMaterial.CODES.COMPOSE_RESOURCES(RawMaterial.MATERIAL_VEGETATION);
 		for(Integer code : codes)
-			if(!CMParms.contains(Chant_SummonSeed.NON_SEEDS,code.intValue()))
+			if(!CMParms.contains(Chant_SummonSeed.NON_SEEDS,code))
 			{
 				choices.addElement(code);
 				String desc=RawMaterial.CODES.NAME(code.intValue());
 				if((s.length()>0)&&(CMLib.english().containsString(desc,s)))
-					material=code;
+					material=code.intValue();
 				if(col==4){ buf.append("\n\r"); col=0;}
 				col++;
 				buf.append(CMStrings.padRight(CMStrings.capitalizeAndLower(desc),15));
