+++ com/planet_ink/coffee_mud/Abilities/Druid/Chant_SummonSeed.java
@@ -59,7 +59,7 @@ public class Chant_SummonSeed extends Chant
 		List<Integer> codes = RawMaterial.CODES.COMPOSE_RESOURCES(RawMaterial.MATERIAL_VEGETATION);
 		for(Integer code : codes)
 		{
-			if(!CMParms.contains(Chant_SummonSeed.NON_SEEDS,code.intValue()))
+			if(!CMParms.contains(Chant_SummonSeed.NON_SEEDS,code))
             {
 				String str=RawMaterial.CODES.NAME(code.intValue());
 				if(str.toUpperCase().equalsIgnoreCase(s))
